requirements.txt
Django==3.1.0
dill
typing-extensions==4.4.0
psycopg2>=2.8.6,<2.9 
pandas~= 1.2.3
numpy==1.19.2
scikit-learn==0.24.2
djangorestframework==3.13.1
uWSGI>=2.0.19.1,<2.1
psutil
django-cors-headers
lightgbm
xgboost






